requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
botocore==1.35.63
cement==2.10.14
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
dj-database-url==2.3.0
Django==5.1.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.16
idna==3.10
jmespath==1.0.1
mysqlclient==2.2.6
pathspec==0.10.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.3
semantic-version==2.10.0
six==1.16.0
sqlparse==0.5.2
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
